+++ src/fhirpath/types.py
@@ -317,7 +317,7 @@ class FhirDate(FhirPrimitiveType):
 
 
 class FhirDateTime(FhirPrimitiveType):
-    """	A date, date-time or partial date (e.g. just year or year + month) as used in
+    """A date, date-time or partial date (e.g. just year or year + month) as used in
     human communication. The format is YYYY, YYYY-MM, YYYY-MM-DD or
     YYYY-MM-DDThh:mm:ss+zz:zz, e.g. 2018, 1973-06, 1905-08-23,
     2015-02-07T13:28:17-05:00 or 2017-01-01T00:00:00.000Z.